requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pgvector==0.3.4

# OpenAI
openai==1.51.2

# Telegram Bot
python-telegram-bot==21.6
aiohttp==3.10.9

# Calendar Integration
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
caldav==1.3.9

# Task Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
httpx==0.27.2
loguru==0.7.2

# Date/Time
python-dateutil==2.9.0

# Security
cryptography==43.0.1

